# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Graphify Knowledge Agent API
  description: Authenticated multi-project API with named SSE conversation streams.
  version: 1.0.0-poc
paths:
...
  /api/v1/conversations:
    post:
      tags:
      - conversations
      summary: Create Conversation
      operationId: create_conversation_api_v1_conversations_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/CreateConversationRequest'
              - type: 'null'
              title: Body
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/projects/{project_id}/conversations:
    get:
      tags:
      - conversations
      summary: List Conversations
      operationId: list_conversations_api_v1_projects__project_id__conversations_get
      security:
      - HTTPBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: state
        in: query
        required: false
        schema:
          enum:
          - active
          - archived
          type: string
          default: active
          title: State
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            minLength: 1
            maxLength: 1024
          - type: 'null'
          title: Cursor
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/conversations/{conversation_id}:
    get:
      tags:
      - conversations
      summary: Get Conversation
      operationId: get_conversation_api_v1_conversations__conversation_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - conversations
      summary: Rename Conversation
      operationId: rename_conversation_api_v1_conversations__conversation_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConversationRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - conversations
      summary: Delete Conversation
      operationId: delete_conversation_api_v1_conversations__conversation_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/conversations/{conversation_id}/restore:
    post:
      tags:
      - conversations
      summary: Restore Conversation
      operationId: restore_conversation_api_v1_conversations__conversation_id__restore_post
      security:
      - HTTPBearer: []
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/conversations/{conversation_id}/purge:
    delete:
      tags:
      - conversations
      summary: Purge Conversation
      operationId: purge_conversation_api_v1_conversations__conversation_id__purge_delete
      security:
      - HTTPBearer: []
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/conversations/{conversation_id}/messages:
    post:
      tags:
      - conversations
      summary: Send Message
      operationId: send_message_api_v1_conversations__conversation_id__messages_post
      security:
      - HTTPBearer: []
      parameters:
      - name: conversation_id
        in: path
        required: true
        schema:
          type: string
          title: Conversation Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Successful Response
          content:
            text/event-stream:
              schema:
                type: string
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Answer:
      properties:
        requestId:
          type: string
          title: Requestid
        conversationId:
          type: string
          title: Conversationid
        answer:
          type: string
          title: Answer
        status:
          type: string
          const: completed
          title: Status
          default: completed
        responseType:
          type: string
          enum:
          - answer
          - clarification
          - insufficient
          title: Responsetype
          default: answer
        confidence:
          type: string
          enum:
          - high
          - medium
          - low
          - insufficient
          title: Confidence
        graphVersion:
          anyOf:
          - type: string
          - type: 'null'
          title: Graphversion
        citations:
          items:
            $ref: '#/components/schemas/Citation'
          type: array
          title: Citations
        graphEvidence:
          $ref: '#/components/schemas/GraphEvidence'
        warnings:
          items:
            type: string
          type: array
          title: Warnings
      additionalProperties: false
      type: object
      required:
      - requestId
      - conversationId
      - answer
      - confidence
      - graphEvidence
      title: Answer
...
    Citation:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        source:
          type: string
          title: Source
        nodeId:
          anyOf:
          - type: string
          - type: 'null'
          title: Nodeid
        relationship:
          anyOf:
          - type: string
          - type: 'null'
          title: Relationship
        provenance:
          type: string
          enum:
          - explicit
          - extracted
          - inferred
          - unknown
          title: Provenance
          default: unknown
        excerpt:
          anyOf:
          - type: string
          - type: 'null'
          title: Excerpt
        document:
          anyOf:
          - type: string
          - type: 'null'
          title: Document
        article:
          anyOf:
          - type: string
          - type: 'null'
          title: Article
        paragraph:
          anyOf:
          - type: string
          - type: 'null'
          title: Paragraph
        startLine:
          anyOf:
          - type: integer
          - type: 'null'
          title: Startline
        endLine:
          anyOf:
          - type: integer
          - type: 'null'
          title: Endline
        pageNumber:
          anyOf:
          - type: integer
            minimum: 1.0
          - type: 'null'
          title: Pagenumber
        sectionPath:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Sectionpath
      additionalProperties: false
      type: object
      required:
      - id
      - title
      - source
      title: Citation
    Conversation:
      properties:
        id:
          type: string
          title: Id
        projectId:
          type: string
          title: Projectid
        name:
          type: string
          title: Name
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
        archivedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Archivedat
        messages:
          items:
            $ref: '#/components/schemas/Message'
          type: array
          title: Messages
      additionalProperties: false
      type: object
      required:
      - id
      - projectId
      - name
      - createdAt
      - updatedAt
      - archivedAt
      title: Conversation
    ConversationList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ConversationSummary'
          type: array
          maxItems: 100
          title: Items
        nextCursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Nextcursor
      additionalProperties: false
      type: object
      required:
      - items
      - nextCursor
      title: ConversationList
    ConversationSummary:
      properties:
        id:
          type: string
          title: Id
        projectId:
          type: string
          title: Projectid
        name:
          type: string
          title: Name
        createdAt:
          type: string
          format: date-time
          title: Createdat
        updatedAt:
          type: string
          format: date-time
          title: Updatedat
        archivedAt:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Archivedat
      additionalProperties: false
      type: object
      required:
      - id
      - projectId
      - name
      - createdAt
      - updatedAt
      - archivedAt
      title: ConversationSummary
...
    CreateConversationRequest:
      properties:
        projectId:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Projectid
      additionalProperties: false
      type: object
      title: CreateConversationRequest
...
    GraphEdge:
      properties:
        id:
          type: string
          title: Id
        sourceNodeId:
          type: string
          title: Sourcenodeid
        targetNodeId:
          type: string
          title: Targetnodeid
        relationship:
          type: string
          title: Relationship
        properties:
          additionalProperties:
            anyOf:
            - type: string
            - type: integer
            - type: number
            - type: boolean
            - type: 'null'
          type: object
          title: Properties
        provenance:
          type: string
          enum:
          - explicit
          - extracted
          - inferred
          - unknown
          title: Provenance
          default: unknown
      type: object
      required:
      - id
      - sourceNodeId
      - targetNodeId
      - relationship
      title: GraphEdge
    GraphEvidence:
      properties:
        nodes:
          items:
            $ref: '#/components/schemas/GraphNode'
          type: array
          title: Nodes
        edges:
          items:
            $ref: '#/components/schemas/GraphEdge'
          type: array
          title: Edges
        paths:
          items:
            $ref: '#/components/schemas/GraphPath'
          type: array
          title: Paths
      type: object
      title: GraphEvidence
    GraphNode:
      properties:
        id:
          type: string
          title: Id
        label:
          type: string
          title: Label
        type:
          type: string
          title: Type
          default: entity
        properties:
          additionalProperties:
            anyOf:
            - type: string
            - type: integer
            - type: number
            - type: boolean
            - type: 'null'
          type: object
          title: Properties
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
        excerpt:
          anyOf:
          - type: string
          - type: 'null'
          title: Excerpt
        provenance:
          type: string
          enum:
          - explicit
          - extracted
          - inferred
          - unknown
          title: Provenance
          default: unknown
      type: object
      required:
      - id
      - label
      title: GraphNode
    GraphPath:
      properties:
        id:
          type: string
          title: Id
        nodeIds:
          items:
            type: string
          type: array
          title: Nodeids
        edgeIds:
          items:
            type: string
          type: array
          title: Edgeids
      type: object
      required:
      - id
      - nodeIds
      title: GraphPath
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Message:
      properties:
        id:
          type: string
          title: Id
        role:
          type: string
          enum:
          - user
          - assistant
          title: Role
        content:
          type: string
          title: Content
        status:
          type: string
          enum:
          - pending
          - completed
          - failed
          title: Status
        createdAt:
          type: string
          format: date-time
          title: Createdat
        result:
          anyOf:
          - $ref: '#/components/schemas/Answer'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - id
      - role
      - content
      - status
      - createdAt
      title: Message
...
    SendMessageRequest:
      properties:
        message:
          type: string
          maxLength: 4000
          minLength: 1
          title: Message
        includeGraphPaths:
          type: boolean
          title: Includegraphpaths
          default: true
      additionalProperties: false
      type: object
      required:
      - message
      title: SendMessageRequest
...
    UpdateConversationRequest:
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 1
          title: Name
      additionalProperties: false
      type: object
      required:
      - name
      title: UpdateConversationRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer